+++ NEWS.md
@@ -34,6 +34,8 @@
     
     - Any `cols` argument in `{openair}` can now take the `colourOpts()` function, which tells each plotting function how to use the new `openColours()` arguments.
     
+    - Added additional palettes by Paul Tol; "tol.highcontrast", "tol.vibrant", "tol.mediumcontrast", "tol.pale" and "tol.dark".
+    
     - Added `openColors()` and `colorOpts()` which are synonymous with their British English equivalents `openColours()` and `colourOpts()`.
 
 - `smoothTrend()` will now use `loess` when it has insufficient data to fit a GAM.